# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "WEMU API Documentation",
        "description": "Swagger documentation for all APIs",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/docs",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/dashboard": {
            "get": {
                "tags": [
                    "Dashboard"
                ],
                "summary": "Get user dashboard",
                "operationId": "d5fcadff7c52501c317a59bc00886f22",
                "responses": {
                    "200": {
                        "description": "Dashboard fetched successfully"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/dashboard/section-details": {
            "get": {
                "tags": [
                    "Dashboard"
                ],
                "summary": "Get section details",
                "operationId": "388ab6e0a1eea5b06bd1d4db814f03a0",
                "parameters": [
                    {
                        "name": "type_id",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Section details fetched successfully"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
}